requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.17.2
altair==4.2.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
appdirs==1.4.4
asyncssh==2.21.1
atpublic==7.0.0
attrs==25.4.0
billiard==4.2.4
black==25.12.0
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.76.0
deprecation==2.1.0
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.24.10
dvc==3.64.2
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
dynaconf==3.2.12
entrypoints==0.4
evidently==0.7.18
exceptiongroup==1.3.1
Faker==39.0.0
fastapi==0.124.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.2
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
great_expectations==1.10.0
gto==1.9.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.5.5
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaggle==1.8.3
kagglesdk==0.1.14
kiwisolver==1.4.9
kombu==5.6.1
librt==0.7.5
lightgbm==4.6.0
litestar==2.19.0
litestar-htmx==0.5.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mistune==3.2.0
mlflow==3.8.0
mlflow-skinny==3.8.0
mlflow-tracing==3.8.0
msgspec==0.20.0
multidict==6.7.0
multipart==1.3.0
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6
nltk==3.9.2
numpy==2.3.5
omegaconf==2.3.0
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optuna==4.6.0
optuna-integration==4.6.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
patsy==1.0.2
pillow==12.0.0
platformdirs==4.5.1
plotly==5.24.1
pluggy==1.6.0
polyfactory==3.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydot==4.0.1
pygit2==1.19.0
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.5
pytest==9.0.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-click==1.9.5
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
scikit-learn==1.7.2
scipy==1.16.3
scmrepo==3.5.9
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.5
sqltrie==0.11.2
starlette==0.50.0
statsmodels==0.14.6
streamlit==1.52.2
tabulate==0.9.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.3
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.1
typer==0.20.0
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.0
uuid6==2025.0.1
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
voluptuous==0.15.2
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.4
xgboost==3.1.2
yarl==1.22.0
zc.lockfile==4.0
zipp==3.23.0
